Removed legacy cuda joint_matrix.
cuda joint_matrix e2e tests refactored so that they will pass on any
tensor cores supporting gpu.
